TTT + TE = OK
