+++ ticTacToeStrategies/st11.py
@@ -1,10 +1,8 @@
-from problems.prob0.game import *
-
-def check(a: GameState, i: int, j: int) -> bool:
+def check(a, i: int, j: int) -> bool:
     return a.a[i][j] == '.'
 
-def Strategy(a: GameState, b: int) -> Turn:
+def Strategy(game, a, b):
    while True:
        i, j = random.randint(0, 2), random.randint(0, 2)
        if check(a, i, j):
-           return Turn(i, j)
+           return game.Turn(i, j)
