+++ trustEvolutionStrategies/change.py
@@ -1,7 +1,5 @@
-from classes import *
-
-def Strategy(a: GameState, b: int) -> Turn:
+def Strategy(game, a, b):
 	if (len(a.turns[0]) == 0):
-		return Turn(0)
+		return game.Turn(0)
 	else:
-		return Turn(a.turns[b ^ 1][-1] ^ 1)
+		return game.Turn(a.turns[b ^ 1][-1] ^ 1)
