+++ trustEvolutionStrategies/random.py
@@ -1,5 +1,4 @@
-from classes import *
 from random import randint
 
-def Strategy(a: GameState, b: int) -> Turn:
-	return Turn(randint(0, 1))	
+def Strategy(game, a, b):
+	return game.Turn(randint(0, 1))
